fix: 폴더 업데이트 오류 수정, 슬라이드로 쿠폰 수정할 때 데이터 못 들고 오는 오류 수정

+++ lib/src/infra/local_db/data_clear_manager.dart
@@ -8,6 +8,8 @@ class DataClearManager {
     await ref.read(allCouponsProvider.notifier).clearAll();
     await ref.read(folderProvider.notifier).clearAll();
     final imageDirectory = await FileHelper.getImagesDirectory();
-    imageDirectory.deleteSync(recursive: true);
+    if (imageDirectory.existsSync()) {
+      imageDirectory.deleteSync(recursive: true);
+    }
   }
 }
